+++ packages/iov-bns/src/decode.ts
@@ -1,4 +1,5 @@
 import {
+  Amount,
   BaseTx,
   FullSignature,
   SendTx,
@@ -8,13 +9,14 @@ import {
   SwapCounterTx,
   SwapIdBytes,
   SwapTimeoutTx,
+  TokenTicker,
   TransactionKind,
   UnsignedTransaction,
 } from "@iov/bcp-types";
 import { ChainId } from "@iov/tendermint-types";
 
 import * as codecImpl from "./codecimpl";
-import { asNumber, decodeFullSig, decodeToken, ensure } from "./types";
+import { asNumber, decodeFullSig, ensure } from "./types";
 import { encodeBnsAddress, isHashIdentifier } from "./util";
 
 // export const buildTx = async (
@@ -28,6 +30,14 @@ import { encodeBnsAddress, isHashIdentifier } from "./util";
 //     signatures: sigs.map(encodeFullSig),
 //   });
 // };
+
+export function decodeAmount(coin: codecImpl.x.ICoin): Amount {
+  return {
+    whole: asNumber(coin.whole),
+    fractional: asNumber(coin.fractional),
+    tokenTicker: (coin.ticker || "") as TokenTicker,
+  };
+}
 
 export const parseTx = (tx: codecImpl.app.ITx, chainId: ChainId): SignedTransaction => {
   const sigs = ensure(tx.signatures, "signatures").map(decodeFullSig);
@@ -58,7 +68,7 @@ const parseSendTx = (base: BaseTx, msg: codecImpl.cash.ISendMsg): SendTx => ({
   //    src: await keyToAddress(tx.signer),
   kind: TransactionKind.Send,
   recipient: encodeBnsAddress(ensure(msg.dest, "recipient")),
-  amount: decodeToken(ensure(msg.amount)),
+  amount: decodeAmount(ensure(msg.amount)),
   memo: msg.memo || undefined,
   ...base,
 });
@@ -79,7 +89,7 @@ const parseSwapCounterTx = (base: BaseTx, msg: codecImpl.escrow.ICreateEscrowMsg
     hashCode,
     recipient: encodeBnsAddress(ensure(msg.recipient, "recipient")),
     timeout: asNumber(msg.timeout),
-    amount: (msg.amount || []).map(decodeToken),
+    amount: (msg.amount || []).map(decodeAmount),
     ...base,
   };
 };
@@ -107,7 +117,7 @@ const parseBaseTx = (tx: codecImpl.app.ITx, sig: FullSignature, chainId: ChainId
     signer: sig.pubkey,
   };
   if (tx.fees && tx.fees.fees) {
-    return { ...base, fee: decodeToken(tx.fees.fees) };
+    return { ...base, fee: decodeAmount(tx.fees.fees) };
   }
   return base;
 };