+++ packages/iov-bns/types/testdata.d.ts
@@ -1,4 +1,4 @@
-import { Address, FullSignature, FungibleToken, SendTx, SignedTransaction } from "@iov/bcp-types";
+import { Address, Amount, FullSignature, SendTx, SignedTransaction } from "@iov/bcp-types";
 import { ChainId, PublicKeyBundle } from "@iov/tendermint-types";
 import { PrivateKeyBundle } from "./types";
 export declare const pubJson: PublicKeyBundle;
@@ -6,7 +6,7 @@ export declare const pubBin: Uint8Array;
 export declare const privJson: PrivateKeyBundle;
 export declare const privBin: Uint8Array;
 export declare const address: Address;
-export declare const coinJson: FungibleToken;
+export declare const coinJson: Amount;
 export declare const coinBin: Uint8Array;
 export declare const chainId: ChainId;
 export declare const sendTxJson: SendTx;