+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 Breaking changes
 
+* @iov/bcp-types: Rename `FungibleToken` to `Amount`
 * @iov/bns: Re-generate BNS codec from weave v0.9.0 and adapt wrapper types.
 * @iov/crypto: Convert pubkey of Secp256k1Keypair into uncompressed format with prefix `04`.
 * @iov/keycontrol: Let `Secp256k1HdWallet` work on uncompressed pubkeys. Since `Secp256k1HdWallet`